refactor: extract the NATS connector into resonate-nats
