+++ packages/resonate-nats/tests/test_layering.py
@@ -1,0 +1,43 @@
+"""resonate-nats depends on resonate-base, never on resonate-sdk.
+
+This is the assertion that makes the base package worth having: a connector
+must be buildable and releasable without the SDK. If this test ever fails, the
+seam has leaked and the next connector will inherit the leak.
+"""
+
+from __future__ import annotations
+
+import ast
+import pathlib
+
+SRC = pathlib.Path(__file__).parent.parent / "src" / "resonate_nats"
+
+FORBIDDEN_ROOTS = {"resonate", "resonate_aws", "resonate_testing"}
+
+
+def modules() -> list[pathlib.Path]:
+    return sorted(SRC.rglob("*.py"))
+
+
+def imported_roots(path: pathlib.Path) -> set[str]:
+    tree = ast.parse(path.read_text(encoding="utf-8"))
+    roots: set[str] = set()
+    for node in ast.walk(tree):
+        if isinstance(node, ast.Import):
+            roots.update(alias.name.split(".")[0] for alias in node.names)
+        elif isinstance(node, ast.ImportFrom) and not node.level and node.module:
+            roots.add(node.module.split(".")[0])
+    return roots
+
+
+def test_source_files_exist() -> None:
+    assert modules(), f"no modules found under {SRC}"
+
+
+def test_connector_does_not_import_the_sdk() -> None:
+    offenders = {
+        path.name: sorted(imported_roots(path) & FORBIDDEN_ROOTS)
+        for path in modules()
+        if imported_roots(path) & FORBIDDEN_ROOTS
+    }
+    assert not offenders, f"the connector reaches into the SDK: {offenders}"
